requirements.txt
absl-py==2.2.2
altair==5.5.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.5.1
gast==0.6.0
gdown==5.2.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.22.2
joblib==1.4.2
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
matplotlib==3.8.0
namex==0.0.9
narwhals==1.40.0
numpy==1.26.1
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
protobuf==4.25.7
pyarrow==20.0.0
pydeck==0.9.1
pyparsing==3.2.3
PySocks==1.7.1
pytz==2025.2
rich==13.9.4
scikit-learn==1.3.1
scipy==1.15.3
seaborn==0.13.2
smmap==5.0.2
streamlit==1.40.2
tenacity==9.1.2
tensorboard-data-server==0.7.2
tensorflow-cpu==2.16.1
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tzdata==2025.2
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
